move treemap into JOverflow module directly
