--- application/org.openjdk.jmc.joverflow.ext.treemap/META-INF/MANIFEST.MF
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-Name: Joverflow Treemap View
-Bundle-SymbolicName: org.openjdk.jmc.joverflow.ext.treemap;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-Vendor: Oracle Corporation
-Require-Bundle: org.openjdk.jmc.ui,
- org.openjdk.jmc.joverflow,
- org.openjdk.jmc.joverflow.ui
-Eclipse-BuddyPolicy: app
-Bundle-ActivationPolicy: lazy
-Bundle-Activator: org.openjdk.jmc.joverflow.ext.treemap.JOverflowTreemapPlugin
-Automatic-Module-Name: org.openjdk.jmc.joverflow.ext.treemap
-Import-Package: org.openjdk.jmc.flightrecorder.ui
